DataHub-Public/gdm/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
Django>=4.1,<4.2
wheel
setuptools
djangorestframework==3.14
markdown       # Markdown support for the browsable API.
django-filter  # Filtering support
pyyaml # generate schema into YAML-based OpenAPI format
uritemplate # used internally to get parameters in path
djangorestframework-jsonapi
# for optional package integrations
# djangorestframework-jsonapi['django-filter']
# djangorestframework-jsonapi['django-polymorphic']
# djangorestframework-jsonapi['openapi']
gunicorn==20.1.0    # python web server gateway for wsgi
gevent>=21,<22
# eventlet==0.30.2    # async web server workers for gunicorn
# uwsgi
psycopg2-binary #>=2.8,<2.9
django-environ
requests
# django-python3-ldap
pandas>=1.5.0,<1.6
apscheduler
Pillow
django-tables2
